+++ lib/dep.ts
@@ -1,10 +1,8 @@
-import { SubscriptionContainer } from './sub';
-import State from './state';
-export default class Dep {
-  // static
+import { State, SubscriptionContainer } from './internal';
+
+export class Dep {
   public deps: Set<any> = new Set();
   public subs: Set<SubscriptionContainer> = new Set();
-  // public dynamic: Set<State> = new Set(); // cleanout foriegn deps on update
 
   constructor(initialDeps?: Array<Dep>) {
     if (initialDeps) initialDeps.forEach(dep => this.deps.add(dep));
@@ -15,3 +13,5 @@ export default class Dep {
     this.deps.add(instance);
   }
 }
+
+export default Dep;